Inline subgroup checks into openvm.rs
Move SubgroupCheck trait and implementations from the separate
openvm_subgroup_check.rs module into the bottom of openvm.rs.
No functional change.
